+++ stream-chat-android-ui-common/src/main/kotlin/com/getstream/sdk/chat/viewmodel/MessageInputViewModel.kt
@@ -23,6 +23,8 @@ import androidx.lifecycle.asLiveData
 import androidx.lifecycle.viewModelScope
 import com.getstream.sdk.chat.utils.extensions.isDirectMessaging
 import com.getstream.sdk.chat.utils.extensions.isModerationFailed
+import com.getstream.sdk.chat.utils.typing.DefaultTypingUpdatesBuffer
+import com.getstream.sdk.chat.utils.typing.TypingUpdatesBuffer
 import io.getstream.chat.android.client.ChatClient
 import io.getstream.chat.android.client.call.enqueue
 import io.getstream.chat.android.client.extensions.cidToTypeAndId
@@ -37,6 +39,7 @@ import io.getstream.chat.android.core.ExperimentalStreamChatApi
 import io.getstream.chat.android.offline.extensions.setMessageForReply
 import io.getstream.chat.android.offline.extensions.watchChannelAsState
 import io.getstream.chat.android.offline.plugin.state.channel.ChannelState
+import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.SharingStarted
 import kotlinx.coroutines.flow.filterNotNull
@@ -54,6 +57,7 @@ import java.io.File
  * @param chatClient Entry point for most of the chat SDK
  * such as the current user, connection state, unread counts etc.
  */
+@OptIn(ExperimentalCoroutinesApi::class)
 @Suppress("TooManyFunctions")
 public class MessageInputViewModel @JvmOverloads constructor(
     private val cid: String,
@@ -113,6 +117,16 @@ public class MessageInputViewModel @JvmOverloads constructor(
             started = SharingStarted.Eagerly,
             initialValue = setOf()
         ).asLiveData()
+
+    /**
+     * Buffers typing updates.
+     *
+     * @see [DefaultTypingUpdatesBuffer]
+     */
+    public var typingUpdatesBuffer: TypingUpdatesBuffer = DefaultTypingUpdatesBuffer(
+        onTypingStarted = ::keystroke,
+        onTypingStopped = ::stopTyping
+    )
 
     /**
      * Holds the message the user is currently replying to,
@@ -388,6 +402,15 @@ public class MessageInputViewModel @JvmOverloads constructor(
         if (repliedMessage.value != null) {
             ChatClient.instance().setMessageForReply(cid, null).enqueue()
         }
+    }
+
+    /**
+     * Performs hygiene events such as clearing typing updates
+     * when the used leaves the messages screen.
+     */
+    override fun onCleared() {
+        super.onCleared()
+        typingUpdatesBuffer.clear()
     }
 
     private companion object {