+++ stream-chat-android-ui-components/src/main/kotlin/io/getstream/chat/android/ui/message/input/MessageInputView.kt
@@ -39,6 +39,7 @@ import com.getstream.sdk.chat.utils.Utils
 import com.getstream.sdk.chat.utils.extensions.activity
 import com.getstream.sdk.chat.utils.extensions.containsLinks
 import com.getstream.sdk.chat.utils.extensions.focusAndShowKeyboard
+import com.getstream.sdk.chat.utils.typing.TypingUpdatesBuffer
 import com.google.android.material.snackbar.Snackbar
 import io.getstream.chat.android.client.logger.ChatLogger
 import io.getstream.chat.android.client.models.Attachment
@@ -111,6 +112,11 @@ public class MessageInputView : ConstraintLayout {
 
     private var onSendButtonClickListener: OnMessageSendButtonClickListener? = null
     private var typingListener: TypingListener? = null
+
+    /**
+     * Used to buffer typing updates in order to conserve API calls.
+     */
+    private var typingUpdatesBuffer: TypingUpdatesBuffer? = null
     private var isKeyboardListenerRegistered: Boolean = false
 
     private var maxMessageLength: Int = Integer.MAX_VALUE
@@ -273,8 +279,27 @@ public class MessageInputView : ConstraintLayout {
         this.onSendButtonClickListener = listener
     }
 
+    @Deprecated(
+        level = DeprecationLevel.WARNING,
+        message = "Use MessageInputView.setTypingListener(TypingUpdatesBuffer) instead to set a buffer " +
+            "which internally manages buffering " +
+            "start typing events and sending stop typing events.\n\n" +
+            "For the default implementation see DefaultTypingUpdatesBuffer.",
+        replaceWith = ReplaceWith(
+            "setTypingUpdatesBuffer(DefaultTypingUpdatesBuffer())",
+            "io.getstream.chat.android.ui.message.input.MessageInputView",
+            "com.getstream.sdk.chat.utils.typing.DefaultTypingUpdatesBuffer"
+        )
+    )
     public fun setTypingListener(listener: TypingListener?) {
         this.typingListener = listener
+    }
+
+    /**
+     * Sets the typing updates buffer.
+     */
+    public fun setTypingUpdatesBuffer(buffer: TypingUpdatesBuffer) {
+        typingUpdatesBuffer = buffer
     }
 
     /**
@@ -871,6 +896,7 @@ public class MessageInputView : ConstraintLayout {
         if (canSendTypingUpdates) {
             if (binding.messageInputFieldView.messageText.isNotEmpty()) {
                 typingListener?.onKeystroke()
+                typingUpdatesBuffer?.onKeystroke()
             } else {
                 typingListener?.onStopTyping()
             }
@@ -1256,6 +1282,17 @@ public class MessageInputView : ConstraintLayout {
         )
     }
 
+    @Deprecated(
+        level = DeprecationLevel.WARNING,
+        message = "Use TypingUpdatesBuffer which internally manages buffering " +
+            "start typing events and sending stop typing events.\n\n" +
+            "For the default implementation see DefaultTypingUpdatesBuffer.",
+        replaceWith = ReplaceWith(
+            "TypingUpdatesBuffer",
+            "com.getstream.sdk.chat.utils.typing.TypingUpdatesBuffer",
+            "com.getstream.sdk.chat.utils.typing.DefaultTypingUpdatesBuffer"
+        )
+    )
     public interface TypingListener {
         public fun onKeystroke()
         public fun onStopTyping()