+++ stream-chat-android-ui-components/src/main/kotlin/io/getstream/chat/android/ui/message/input/viewmodel/MessageInputViewModelBinding.kt
@@ -145,6 +145,8 @@ public fun MessageInputViewModel.bindView(
         }
     )
 
+    view.setTypingUpdatesBuffer(typingUpdatesBuffer)
+
     repliedMessage.observe(lifecycleOwner) {
         if (it != null) {
             view.inputMode = MessageInputView.InputMode.Reply(it)